- Reenabled the Neural/Behavioral alignment code, debugged it, and got it working in a fairly preliminary state.  There is still some work to do (ie. move all database calls to the sessioninfo object) but it is working. - Start/End trial event codes were ints and passing the 7 bits (ie. they went to 128 and beyond) that are sent to the neural data collection device.  I fixed this. - The Start/End trial event codes were being sent to the server before they were sent to the neural data collection device.  This means that the neural event wasn't sent until after the pictoserver responded, which was causing neural/behavioral alignment jitter of ~30ms.  After switching the order, the alignment jitter is submillisecond (usually ~100us). - Added some #ifdefs to get the PictoBoxXPEventCodeGenerator to be usable on an old Orion machine for debugging purposes.  Ideally, I should build a new code generator object for this purpose at some point. - The PictoBoxXPEventCodeGenerator had runtime errors even when running on picto box due to the DAQmxWriteDigitalLines() call that was being used to control the DAQ data lines.  I couldn't find good documentation for this call, and simply went with the tried and true DAQmxWriteDigitalU16 (or DAQmxWriteDigitalU8 in this case) call that is used in Orion.  The U16 call is working on old orion machines and Pictobox no longer has a runtime error, but I haven't yet tried hooking up pictobox to a plexon machine. - Since event code generation now seems to be working, I enabled it in main.cpp of the director. - Small fix for syntaxHighlighter.cpp

+++ source/common/statemachine/StateMachine.h
@@ -109,7 +109,7 @@ private:
 
 	QStringList levelEnumStrs_;
 
-	static int trialEventCode_;
+	static short trialEventCode_;
 	static int trialNum_;
 
 	QSharedPointer<StateMachineElement> currElement_;
